+++ src/ACEtk/interpretation/ContinuousEnergyNeutron/Builder/Tabulated.hpp
@@ -1,8 +1,9 @@
 template< typename T >
 struct Tabulated {
-  std::optional< std::pair<
-      std::vector< double >,
-      std::vector< double > > > parameters;
+  std::optional< 
+    std::pair<
+      std::vector< int >,
+      std::vector< int > > > parameters;
   std::vector< double > x;
   std::vector< T > y;
