+++ src/ACEtk/interpretation/ContinuousEnergyNeutron/Builder/test/Builder.test.cpp
@@ -237,8 +237,8 @@ SCENARIO( "fissionMultiplicity components ofContinuousEnergyNeutron::Builder" ){
         }
       }
       GIVEN( "tabulated fission multiplicity parameters: " + nuType ){
-        std::vector< double > boundaries{ 0.0, 3.0 };
-        std::vector< double > schemes{ 2.0, 1.0 };
+        std::vector< int > boundaries{ 0, 3 };
+        std::vector< int > schemes{ 2, 1 };
         std::vector< double > energies{ 1.0, 2.0, 5.0, 6.0 };
         std::vector< double > multiplicities{ 2.1, 2.2, 2.5, 2.5 };
         CHECK_NOTHROW( 
@@ -269,8 +269,8 @@ SCENARIO( "fissionMultiplicity components ofContinuousEnergyNeutron::Builder" ){
 
 SCENARIO( "PhotonProductionReaction components of ContinuousEnergyNeutron::Builder" ){
   ContinuousEnergyNeutron::Builder ncBuilder{};
-  std::vector< double > boundaries{ 0.0, 3.0 };
-  std::vector< double > schemes{ 2.0, 1.0 };
+  std::vector< int > boundaries{ 0, 3 };
+  std::vector< int > schemes{ 2, 1 };
   std::vector< double > energies{ 1.0, 2.0, 5.0, 6.0 };
   std::vector< double > values{ 2.1, 2.2, 2.5, 2.5 };
